Fix attack.py circular import on boot
backend/api/attack.py imported get_time_range from
backend.api.analytics, but the package's __init__.py imports both
modules — the cross-import re-entered __init__.py while attack.py was
still mid-load and tripped a partial-init ImportError, blocking the
backend from starting.

Move get_time_range into services/mitre_lookup.py (where the other
shared MITRE helpers already live) so backend.api modules don't reach
back into each other through the package. analytics.py keeps its
public name via a re-export so existing imports continue to work.
